nog bezig

+++ api/data/dtos/ingredientsCalculateCASDTO.cs
@@ -6,7 +6,7 @@ namespace api.data.dtos;
 
         public int Id {get; set;}
         public float Height { get; set; }
-        public int Age{ get; set; }
+        public float Age{ get; set; }
         public int Gender {get; set;}
         public int Race {get; set;}
         public float measured {get; set;}
